k8s/apis: refactor CRD registration helpers into a separate package
Factor out the helper functions used to create and install CRD objects
with k8s.
